requirements.txt
# Core Dependencies
requests==2.32.4
python-dotenv==1.0.0

# Google APIs
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# Excel Generation
openpyxl==3.1.2

# Utilities
tqdm==4.66.1
retry==0.9.2

# AI Providers (for individual reports)
google-generativeai==0.8.3  # Gemini AI (default, free tier)

# PDF Generation (for individual reports)
reportlab==4.0.9            # PDF creation
matplotlib==3.8.2           # Charts and visualizations
Pillow==10.2.0              # Image handling for PDFs
